fix: move all stirng utils to shared ui utils
